requirements.txt
Flask==2.3.2
Pillow==9.5.0
ultralytics==8.0.0
numpy==1.25.2
scipy==1.11.2
opencv-python==4.8.0.76
natsort==8.4.0